reduce number of api calls when chaning list names
Originally every character typed in the todo list
name area would cause an api call to be made to
update the name of the todo list. Now api calls to
update the todo list name only happen once the
user clicks off the text box and then an api call
is made to update the todo list name.